Split large stac test into separate tests, module
