+++ integration_tests/conftest.py
@@ -1,32 +1,20 @@
-import pytest
-from pathlib import Path
-
 from click.testing import CliRunner
 
-from cubedash import logs, generate
-from cubedash.summary import SummaryStore
-from datacube.index import Index
-from digitalearthau.testing import factories
-import json
-import re
-from pathlib import Path
-from typing import Dict, Tuple
-
 import pytest
-from click.testing import Result
-from dateutil import tz
-from flask import Response
+from click.testing import CliRunner
 from flask.testing import FlaskClient
+from pathlib import Path
+from typing import Tuple
 
 import cubedash
-from cubedash import _model, _monitoring, _pages
-from cubedash.summary import SummaryStore, show
-from cubedash.summary import _extents
+from cubedash import _model
+from cubedash import logs, generate
+from cubedash.summary import SummaryStore
 from datacube.index import Index
 from datacube.index.hl import Doc2Dataset
 from datacube.model import Dataset
 from datacube.utils import read_documents
-from requests_html import HTML
+from digitalearthau.testing import factories
 
 pytest_plugins = "digitalearthau.testing.plugin"
 
@@ -38,6 +26,8 @@ pytest_plugins = "digitalearthau.testing.plugin"
 module_db = factories.db_fixture('local_config', scope='module')
 module_index = factories.index_fixture('module_db', scope='module')
 module_dea_index = factories.dea_index_fixture('module_index', scope='module')
+
+TEST_DATA_DIR = Path(__file__).parent / 'data'
 
 
 @pytest.fixture(scope='function')
@@ -140,3 +130,33 @@ def client(unpopulated_client: FlaskClient) -> FlaskClient:
         _model.STORE.get_or_update(product.name)
     return unpopulated_client
 
+
+@pytest.fixture(scope='module')
+def populated_index(dataset_loader, module_dea_index):
+    """
+    Index populated with example datasets. Assumes our tests wont modify the data!
+
+    It's module-scoped as it's expensive to populate.
+    """
+    loaded = dataset_loader(
+        'wofs_albers',
+        TEST_DATA_DIR / 'wofs-albers-sample.yaml.gz'
+    )
+    assert loaded == 11
+
+    loaded = dataset_loader(
+        'high_tide_comp_20p',
+        TEST_DATA_DIR / 'high_tide_comp_20p.yaml.gz'
+    )
+    assert loaded == 306
+
+    # These have very large footprints, as they were unioned from many almost-identical
+    # polygons and not simplified. They will trip up postgis if used naively.
+    # (postgis gist index has max record size of 8k per entry)
+    loaded = dataset_loader(
+        'pq_count_summary',
+        TEST_DATA_DIR / 'pq_count_summary.yaml.gz'
+    )
+    assert loaded == 20
+
+    return module_dea_index
