+++ integration_tests/schemas/stac/item.json
