+++ integration_tests/test_page_loads.py
@@ -2,16 +2,12 @@
 Tests that load pages and check the contained text.
 """
 import json
-import jsonschema
 import pytest
 from click.testing import Result
 from dateutil import tz
 from flask import Response
 from flask.testing import FlaskClient
-from pathlib import Path
-from typing import Dict, Optional
 
-import cubedash._stac
 from cubedash import _monitoring, _model
 from cubedash.summary import SummaryStore, show
 from cubedash.summary import _extents
@@ -19,40 +15,15 @@ from datacube.index import Index
 from integration_tests.asserts import get_html, check_dataset_count, \
     check_last_processed, check_area, get_geojson, get_json
 
-TEST_DATA_DIR = Path(__file__).parent / 'data'
-
-
 DEFAULT_TZ = tz.gettz('Australia/Darwin')
 
+
 @pytest.fixture(scope='module', autouse=True)
-def populate_index(dataset_loader, module_dea_index):
+def auto_populate_index(populated_index):
     """
-    Index populated with example datasets. Assumes our tests wont modify the data!
-
-    It's module-scoped as it's expensive to populate.
+    Auto-populate the index for all tests in this file.
     """
-    loaded = dataset_loader(
-        'wofs_albers',
-        TEST_DATA_DIR / 'wofs-albers-sample.yaml.gz'
-    )
-    assert loaded == 11
-
-    loaded = dataset_loader(
-        'high_tide_comp_20p',
-        TEST_DATA_DIR / 'high_tide_comp_20p.yaml.gz'
-    )
-    assert loaded == 306
-
-    # These have very large footprints, as they were unioned from many almost-identical
-    # polygons and not simplified. They will trip up postgis if used naively.
-    # (postgis gist index has max record size of 8k per entry)
-    loaded = dataset_loader(
-        'pq_count_summary',
-        TEST_DATA_DIR / 'pq_count_summary.yaml.gz'
-    )
-    assert loaded == 20
-
-    return module_dea_index
+    return populated_index
 
 
 def test_default_redirect(client: FlaskClient):
@@ -446,137 +417,3 @@ def test_plain_product_list(client: FlaskClient):
     rv: Response = client.get('/products.txt')
     assert 'ls7_nbar_scene\n' in rv.data.decode('utf-8')
 
-
-def test_stac_search(client: FlaskClient):
-    cubedash._stac.DATASET_LIMIT = DATASET_LIMIT = 20
-    cubedash._stac.DEFAULT_PAGE_SIZE = DEFAULT_PAGE_SIZE = 4
-
-    # Search all products
-    limit = DEFAULT_PAGE_SIZE // 2
-    geojson = get_geojson(
-        client,
-        (
-            f'/stac/search?'
-            f'&bbox=[114, -33, 153, -10]'
-            f'&time=2017-04-16T01:12:16/2017-05-10T00:24:21'
-            f'&limit={limit}'
-        )
-    )
-    assert len(geojson.get('features')) == limit
-    dataset_count = limit
-
-    # Keep loading "next" pages and we should hit the DATASET_LIMIT.
-    next_page_url = _get_next_href(geojson)
-    while next_page_url:
-        geojson = get_geojson(client, next_page_url)
-        assert len(geojson.get('features')) == limit
-        dataset_count += limit
-        next_page_url = _get_next_href(geojson)
-    assert dataset_count == DATASET_LIMIT
-
-    # Limit with value greater than DATASET_LIMIT should be truncated.
-    geojson = get_geojson(
-        client,
-        (
-            f'/stac/search?'
-            f'&bbox=[114, -33, 153, -10]'
-            f'&time=2017-04-16T01:12:16/2017-05-10T00:24:21'
-            f'&limit={DATASET_LIMIT + 2}'
-        )
-    )
-    assert len(geojson.get('features')) == DATASET_LIMIT
-
-    # Without limit, it should use the default page size
-    geojson = get_geojson(
-        client,
-        (
-            '/stac/search?'
-            '&bbox=[114, -33, 153, -10]'
-            '&time=2017-04-16T01:12:16/2017-05-10T00:24:21'
-        )
-    )
-    assert len(geojson.get('features')) == DEFAULT_PAGE_SIZE
-
-    # Outside the box there should be no results
-    geojson = get_geojson(
-        client,
-        (
-            '/stac/search?'
-            '&bbox=[20,-5,25,10]'
-            '&time=2017-04-16T01:12:16/2017-05-10T00:24:21'
-        )
-    )
-    assert len(geojson.get('features')) == 0
-
-    # Search a whole-day for a scene
-    geojson = get_geojson(
-        client,
-        (
-            '/stac/search?'
-            'product=ls7_nbar_scene'
-            '&bbox=[114, -33, 153, -10]'
-            '&time=2017-04-20'
-        )
-    )
-    assert len(geojson.get('features')) == 1
-
-    # Search a whole-day on an empty day.
-    geojson = get_geojson(
-        client,
-        (
-            '/stac/search?'
-            'product=ls7_nbar_scene'
-            '&bbox=[114, -33, 153, -10]'
-            '&time=2017-04-22'
-        )
-    )
-    assert len(geojson.get('features')) == 0
-
-    # Test POST, product, and assets
-    rv: Response = client.post(
-        '/stac/search',
-        data=json.dumps({
-            'product': 'wofs_albers',
-            'bbox': [114, -33, 153, -10],
-            'time': '2017-04-16T01:12:16/2017-05-10T00:24:21',
-            'limit': DEFAULT_PAGE_SIZE,
-        }),
-        headers={'Content-Type': 'application/json',
-                 'Accept': 'application/json'}
-    )
-    geodata = json.loads(rv.data)
-    assert len(geodata.get('features')) == DEFAULT_PAGE_SIZE
-    assert 'water' in geodata['features'][0]['assets']
-    assert len(geodata['features'][0]['assets']) == 1
-
-    # Test high_tide_comp_20p with POST and assets
-    rv: Response = client.post(
-        '/stac/search',
-        data=json.dumps({
-            'product': 'high_tide_comp_20p',
-            'bbox': [114, -40, 147, -32],
-            'time': '2000-01-01T00:00:00/2016-10-31T00:00:00',
-            'limit': 5,
-        }),
-        headers={'Content-Type': 'application/json',
-                 'Accept': 'application/json'}
-    )
-    geodata = json.loads(rv.data)
-    bands = ['blue', 'green', 'nir', 'red', 'swir1', 'swir2']
-    for band in bands:
-        assert band in geodata['features'][0]['assets']
-
-    # Validate stac items with jsonschema
-    with open(Path(__file__).parent / 'schemas' / 'stac_item.json', 'r') as fp:
-        schema = json.load(fp)
-    jsonschema.validate(geodata['features'][0], schema)
-
-
-def _get_next_href(geojson: Dict) -> Optional[str]:
-    hrefs = [link['href'] for link in geojson.get('links', []) if link['rel'] == 'next']
-    if not hrefs:
-        return None
-
-    assert len(hrefs) == 1, "Multiple next links found: " + ",".join(hrefs)
-    [href] = hrefs
-    return href
